subzero pre-integration refactor
* move common types/functions used in sql_over_http to errors.rs and http_util.rs
* add the "Local" auth backend to proxy (similar to local_proxy), useful in local testing
* change the Connect and Send type for the http client to allow for custom body when making post requests to local_proxy from the proxy
